+++ example/lib/page_views.dart
@@ -4,33 +4,33 @@ import 'package:flutter/material.dart';
 
 class ViewsPage extends StatelessWidget {
   final List<String> viewNames = ['viewName', 'viewName1'];
-  final List<String> viewIDs = [];
+  final List<String> viewIDs = ['temp_id_1', 'temp_id_2'];//set initial temporary values
 
-  void recordViewHome() {
+  Future<void> recordViewHome() async {
     Map<String, Object> segments = {'Cats': 123, 'Moons': 9.98, 'Moose': 'Deer'};
     // ignore: deprecated_member_use
-    Countly.recordView('HomePage', segments);
+    await Countly.recordView('HomePage', segments);
   }
 
-  void recordViewDashboard() {
+  Future<void> recordViewDashboard() async {
     // ignore: deprecated_member_use
-    Countly.recordView('Dashboard');
+    await Countly.recordView('Dashboard');
   }
 
-  void stopViewWithID() {
-    Countly.instance.views.stopViewWithID(viewIDs[0]);
+  Future<void> stopViewWithID() async {
+    await Countly.instance.views.stopViewWithID(viewIDs[0]);
   }
 
-  void stopViewWithName() {
-    Countly.instance.views.stopViewWithName(viewNames[0]);
+  Future<void> stopViewWithName() async {
+    await Countly.instance.views.stopViewWithName(viewNames[0]);
   }
 
-  void pauseViewWithID() {
-    Countly.instance.views.pauseViewWithID(viewIDs[0]);
+  Future<void> pauseViewWithID() async {
+    await Countly.instance.views.pauseViewWithID(viewIDs[0]);
   }
 
-  void resumeViewWithID() {
-    Countly.instance.views.resumeViewWithID(viewIDs[0]);
+  Future<void> resumeViewWithID() async {
+    await Countly.instance.views.resumeViewWithID(viewIDs[0]);
   }
 
   Future<void> startViewWithSegmentation() async {
@@ -43,12 +43,20 @@ class ViewsPage extends StatelessWidget {
     print(viewIDs[1]);
   }
 
-  void setGlobalViewSegmentation() {
-    Countly.instance.views.setGlobalViewSegmentation({'abcd': '123'});
+  Future<void> setGlobalViewSegmentation() async {
+    await Countly.instance.views.setGlobalViewSegmentation({'abcd': '123'});
   }
 
-  void updateGlobalViewSegmentation() {
-    Countly.instance.views.updateGlobalViewSegmentation({'abcd': '123'});
+  Future<void> updateGlobalViewSegmentation() async {
+    await Countly.instance.views.updateGlobalViewSegmentation({'abcd': '123'});
+  }
+
+  Future<void> addSegmentationToViewWithID() async {
+    await Countly.instance.views.addSegmentationToViewWithID(viewIDs[0], {'aaa': '111'});
+  }
+
+  Future<void> addSegmentationToViewWithName() async {
+    await Countly.instance.views.addSegmentationToViewWithName(viewNames[0], {'bbb': '222'});
   }
 
   @override
@@ -60,20 +68,23 @@ class ViewsPage extends StatelessWidget {
       body: SingleChildScrollView(
         padding: EdgeInsets.all(15),
         child: Center(
-            child: Column(
-          children: [
-            MyButton(text: "Record View: 'HomePage'", color: 'olive', onPressed: recordViewHome),
-            MyButton(text: "Record View: 'Dashboard'", color: 'olive', onPressed: recordViewDashboard),
-            MyButton(text: 'Start View', color: 'yellow', onPressed: startView),
-            MyButton(text: 'Start View With Segmentation', color: 'yellow', onPressed: startViewWithSegmentation),
-            MyButton(text: 'Stop View with ID', color: 'orange', onPressed: stopViewWithID),
-            MyButton(text: 'Stop View with Name', color: 'orange', onPressed: stopViewWithName),
-            MyButton(text: 'Pause View with ID', color: 'yellow', onPressed: pauseViewWithID),
-            MyButton(text: 'Resume View with ID', color: 'yellow', onPressed: resumeViewWithID),
-            MyButton(text: 'Set Global View Segmentation', color: 'grey', onPressed: setGlobalViewSegmentation),
-            MyButton(text: 'Update Global View Segmentation', color: 'grey', onPressed: updateGlobalViewSegmentation),
-          ],
-        )),
+          child: Column(
+            children: [
+              MyButton(text: "Record View: 'HomePage'", color: 'olive', onPressed: recordViewHome),
+              MyButton(text: "Record View: 'Dashboard'", color: 'olive', onPressed: recordViewDashboard),
+              MyButton(text: 'Start View', color: 'yellow', onPressed: startView),
+              MyButton(text: 'Start View With Segmentation', color: 'yellow', onPressed: startViewWithSegmentation),
+              MyButton(text: 'Stop View with ID', color: 'orange', onPressed: stopViewWithID),
+              MyButton(text: 'Stop View with Name', color: 'orange', onPressed: stopViewWithName),
+              MyButton(text: 'Pause View with ID', color: 'yellow', onPressed: pauseViewWithID),
+              MyButton(text: 'Resume View with ID', color: 'yellow', onPressed: resumeViewWithID),
+              MyButton(text: 'Set Global View Segmentation', color: 'grey', onPressed: setGlobalViewSegmentation),
+              MyButton(text: 'Update Global View Segmentation', color: 'grey', onPressed: updateGlobalViewSegmentation),
+              MyButton(text: 'Add Segmentation to View with ID', color: 'grey', onPressed: addSegmentationToViewWithID),
+              MyButton(text: 'Add Segmentation to View with Name', color: 'grey', onPressed: addSegmentationToViewWithName),
+            ],
+          ),
+        ),
       ),
     );
   }