+++ ios/Classes/CountlyiOS/CountlyViewTrackingInternal.m
@@ -98,7 +98,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
         
         self.viewDataDictionary = NSMutableDictionary.new;
         self.viewSegmentation = nil;
-        self.isFirstView = true;
+        self.isFirstView = YES;
     }
     
     return self;
@@ -167,7 +167,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
         return nil;
     }
 #endif
-    NSString* viewID = [self startViewInternal:viewName customSegmentation:segmentation isAutoStoppedView:true];
+    NSString* viewID = [self startViewInternal:viewName customSegmentation:segmentation isAutoStoppedView:YES];
     return viewID;
 }
 
@@ -329,7 +329,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 
 - (void)stopViewWithIDInternal:(NSString *) viewKey customSegmentation:(NSDictionary *)customSegmentation
 {
-    [self stopViewWithIDInternal:viewKey customSegmentation:customSegmentation autoPaused:false];
+    [self stopViewWithIDInternal:viewKey customSegmentation:customSegmentation autoPaused:NO];
 }
 
 - (void)stopViewWithIDInternal:(NSString *) viewKey customSegmentation:(NSDictionary *)customSegmentation autoPaused:(BOOL) autoPaused{
@@ -347,6 +347,11 @@ NSString* const kCountlyVTKeyDur      = @"dur";
         NSMutableDictionary* segmentation = NSMutableDictionary.new;
         segmentation[kCountlyVTKeyName] = viewData.viewName;
         segmentation[kCountlyVTKeySegment] = CountlyDeviceInfo.osName;
+        
+        if (viewData.segmentation)
+        {
+            [segmentation addEntriesFromDictionary:viewData.segmentation];
+        }
         
         if (self.viewSegmentation)
         {
@@ -375,7 +380,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 
 - (NSString*)startViewInternal:(NSString *)viewName customSegmentation:(NSDictionary *)customSegmentation
 {
-    return [self startViewInternal:viewName customSegmentation:customSegmentation isAutoStoppedView:false];
+    return [self startViewInternal:viewName customSegmentation:customSegmentation isAutoStoppedView:NO];
 }
 
 - (NSString*)startViewInternal:(NSString *)viewName customSegmentation:(NSDictionary *)customSegmentation isAutoStoppedView:(BOOL) isAutoStoppedView
@@ -404,7 +409,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     
     if (self.isFirstView)
     {
-        self.isFirstView = false;
+        self.isFirstView = NO;
         segmentation[kCountlyVTKeyStart] = @1;
     }
     
@@ -501,29 +506,17 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     }
 }
 
-- (void)pauseCurrentView
-{
-    if (self.currentView)
-    {
-        [self pauseViewInternal:self.currentView];
-    }
-}
-
-- (void)resumeCurrentView
-{
-    [self.currentView resumeView];
-}
 
 - (void)pauseAllViewsInternal
 {
     [self.viewDataDictionary enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull key, CountlyViewData * _Nonnull viewData, BOOL * _Nonnull stop) {
-        [self pauseViewInternal:viewData autoPaused:true];
+        [self pauseViewInternal:viewData autoPaused:YES];
     }];
 }
 
 - (void)pauseViewInternal:(CountlyViewData*) viewData
 {
-    [self pauseViewInternal:viewData];
+    [self pauseViewInternal:viewData autoPaused:NO];
 }
 
 - (void)pauseViewInternal:(CountlyViewData*) viewData autoPaused:(BOOL) autoPaused
@@ -534,7 +527,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     else {
         [viewData pauseView];
     }
-    [self stopViewWithIDInternal:viewData.viewID customSegmentation:nil autoPaused:true];
+    [self stopViewWithIDInternal:viewData.viewID customSegmentation:nil autoPaused:YES];
 }
 
 - (void)resumeAllViewsInternal
@@ -556,6 +549,73 @@ NSString* const kCountlyVTKeyDur      = @"dur";
     }];
 }
 
+- (void)addSegmentationToViewWithNameInternal:(NSString *) viewName segmentation:(NSDictionary *)segmentation
+{
+    if (!viewName || !viewName.length)
+    {
+        CLY_LOG_D(@"%s View name should not be null or empty", __FUNCTION__);
+        return;
+    }
+    
+    __block NSString *viewID = nil;
+    [self.viewDataDictionary enumerateKeysAndObjectsUsingBlock:^(NSString * key, CountlyViewData * viewData, BOOL * stop)
+     {
+        if ([viewData.viewName isEqualToString:viewName])
+        {
+            viewID = key;
+            *stop = YES;
+        }
+        
+    }];
+    
+    if (viewID)
+    {
+        [self addSegmentationToViewWithIDInternal:viewID segmentation:segmentation];
+    }
+    else {
+        CLY_LOG_D(@"%s No View exist with name: %@", __FUNCTION__, viewName);
+    }
+}
+
+- (void)addSegmentationToViewWithIDInternal:(NSString *) viewID segmentation:(NSDictionary *)segmentation{
+    if (!viewID || !viewID.length)
+    {
+        CLY_LOG_D(@"%s View ID should not be null or empty", __FUNCTION__);
+        return;
+    }
+    
+    if (!CountlyConsentManager.sharedInstance.consentForViewTracking)
+        return;
+    CountlyViewData* viewData = self.viewDataDictionary[viewID];
+    if (viewData)
+    {
+        NSMutableDictionary *mutableSegmentation = segmentation.mutableCopy;
+        [mutableSegmentation removeObjectsForKeys:self.reservedViewTrackingSegmentationKeys];
+        if(mutableSegmentation) {
+            if(!viewData.segmentation) {
+                viewData.segmentation = NSMutableDictionary.new;
+            }
+            [viewData.segmentation addEntriesFromDictionary:mutableSegmentation];
+        }
+        [self.viewDataDictionary setObject:viewData forKey:viewID];
+    }
+    else {
+        CLY_LOG_D(@"%s No View exist with ID: %@", __FUNCTION__, viewID);
+    }
+}
+
+
+- (void)addSegmentationToViewWithID:(NSString *)viewID segmentation:(NSDictionary *)segmentation
+{
+    CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewID, segmentation);
+    [self addSegmentationToViewWithIDInternal:viewID segmentation:segmentation];
+}
+
+- (void)addSegmentationToViewWithName:(NSString *)viewName segmentation:(NSDictionary *)segmentation
+{
+    CLY_LOG_I(@"%s %@ %@", __FUNCTION__, viewName, segmentation);
+    [self addSegmentationToViewWithNameInternal:viewName segmentation:segmentation];
+}
 
 #pragma mark - Internal auto view tracking methods
 
@@ -674,7 +734,7 @@ NSString* const kCountlyVTKeyDur      = @"dur";
 
 - (void)resetFirstView
 {
-    self.isFirstView = false;
+    self.isFirstView = NO;
 }