+++ ios/Classes/CountlyiOS/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 23.12.0
+* Added `disableLocation` initial config property to disable location tracking
+* Added `addSegmentationToViewWithID:` in view interface for adding segmentation to an ongoing view
+* Added `addSegmentationToViewWithName:` in view interface for adding segmentation to an ongoing view
+
+* Fixed bug with "pauseViewWithID" call where it could go into a recursive loop
+
 ## 23.8.3
 * Added `requestDropAgeHours` initial config property to set a time limit after which the requests would be removed if not sent to the server
 * Added a call to enroll users to A/B tests when getting a remote config value: 'getValueAndEnroll'